Fix for report log issue
