+++ src/main/java/com/divudi/bean/common/ReportTimerController.java
@@ -3,7 +3,7 @@ package com.divudi.bean.common;
 import com.divudi.core.data.reports.IReportType;
 import com.divudi.core.entity.WebUser;
 import com.divudi.core.entity.report.ReportLog;
-import com.divudi.core.facade.ReportLogFacade;
+import com.divudi.service.ReportLogAsyncService;
 
 import javax.ejb.EJB;
 import javax.ejb.Stateless;
@@ -18,7 +18,7 @@ public class ReportTimerController implements Serializable {
     private static final Logger LOGGER = Logger.getLogger(ReportTimerController.class.getName());
 
     @EJB
-    private ReportLogFacade reportLogFacade;
+    private ReportLogAsyncService reportLogAsyncService;
 
     public void trackReportExecution(Runnable reportGenerationLogic, IReportType reportType, WebUser loggedUser) {
         final Date startTime = new Date();
@@ -33,26 +33,6 @@ public class ReportTimerController implements Serializable {
 
         final ReportLog reportLog = new ReportLog(reportType, loggedUser, startTime, endTime);
 
-        save(reportLog);
-    }
-
-    public void save(ReportLog reportLog) {
-        if (reportLog == null) {
-            return;
-        }
-
-        if (reportLog.getId() == null) {
-            try {
-                getFacade().create(reportLog);
-            } catch (Exception e) {
-                LOGGER.log(Level.SEVERE, "Error occurred while saving the report log", e);
-            }
-        } else {
-            getFacade().edit(reportLog);
-        }
-    }
-
-    public ReportLogFacade getFacade() {
-        return reportLogFacade;
+        reportLogAsyncService.logReport(reportLog);
     }
 }
